+++ init-scripts/03-basin-tables.sql
@@ -23,7 +23,7 @@ CREATE TABLE IF NOT EXISTS basin_watershed_7days (
     water_balance NUMERIC(15,6),
     drought_index INTEGER,
     runoff_index  INTEGER,
-    wb_level      INTEGER
+    wb_level      NUMERIC(15,6)
 );
 
 CREATE INDEX IF NOT EXISTS idx_bw7_date_sim ON basin_watershed_7days(date_sim);
@@ -43,7 +43,7 @@ CREATE TABLE IF NOT EXISTS basin_watershed_6months (
     water_balance NUMERIC(15,6),
     drought_index INTEGER,
     runoff_index  INTEGER,
-    wb_level      INTEGER
+    wb_level      NUMERIC(15,6)
 );
 
 CREATE INDEX IF NOT EXISTS idx_bw6_date_sim ON basin_watershed_6months(date_sim);
@@ -69,7 +69,7 @@ CREATE TABLE IF NOT EXISTS basin_subbasin_l1_7days (
     water_balance NUMERIC(15,6),
     drought_index INTEGER,
     runoff_index  INTEGER,
-    wb_level      INTEGER
+    wb_level      NUMERIC(15,6)
 );
 
 CREATE INDEX IF NOT EXISTS idx_bl17_date_sim ON basin_subbasin_l1_7days(date_sim);
@@ -92,7 +92,7 @@ CREATE TABLE IF NOT EXISTS basin_subbasin_l1_6months (
     water_balance NUMERIC(15,6),
     drought_index INTEGER,
     runoff_index  INTEGER,
-    wb_level      INTEGER
+    wb_level      NUMERIC(15,6)
 );
 
 CREATE INDEX IF NOT EXISTS idx_bl16_date_sim ON basin_subbasin_l1_6months(date_sim);
@@ -118,7 +118,7 @@ CREATE TABLE IF NOT EXISTS basin_subbasin_l2_7days (
     water_balance NUMERIC(15,6),
     drought_index INTEGER,
     runoff_index  INTEGER,
-    wb_level      INTEGER
+    wb_level      NUMERIC(15,6)
 );
 
 CREATE INDEX IF NOT EXISTS idx_bl27_date_sim ON basin_subbasin_l2_7days(date_sim);
@@ -140,7 +140,7 @@ CREATE TABLE IF NOT EXISTS basin_subbasin_l2_6months (
     water_balance NUMERIC(15,6),
     drought_index INTEGER,
     runoff_index  INTEGER,
-    wb_level      INTEGER
+    wb_level      NUMERIC(15,6)
 );
 
 CREATE INDEX IF NOT EXISTS idx_bl26_date_sim ON basin_subbasin_l2_6months(date_sim);
